requirements.txt
openai>=1.40.0
anthropic>=0.39.0
google-generativeai>=0.8.0
lark>=1.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
jinja2>=3.0.0
websockets>=12.0
python-socketio>=5.11.0
python-dotenv>=1.0.0

# Optional but recommended
aiohttp>=3.9.0  # For async HTTP requests
pydantic>=2.5.0  # For data validation
tenacity>=8.2.0  # For retry logic

# Database dependencies
sqlmodel>=0.0.16  # SQLModel for type-safe database models
sqlalchemy>=2.0.0  # SQLAlchemy for database ORM
aiosqlite>=0.19.0  # Async SQLite support

# Container management dependencies
docker>=7.0.0  # Docker SDK for Python

# Git operations dependencies
GitPython>=3.1.0  # Git repository operations
